Add CompanyNavLink component and update package dependencies
