+++ app/routes/admin+/$companyId+/_layout/components/index.ts
@@ -1,0 +1,1 @@
+export * from './CompanyNavLink'
